Add initial perf parser and report generator modules
Two new tools for working with UEFI performance.

- `fpdt_parser`
- `perf_report_generator`

This code is based on the original tools not available on
edk2-pytool-extensions but heavily refactored to improve style,
some code patterns, and other areas like doc strings. Any significant
functional changes from the original code are broken out into
separate focused commits. The tool versions reflect a major version
change to the previously used version to provide continuity for any
previous users.

Documentation is initial at this point and scoped to how the tools
work not UEFI performance in general.

Unit tests were not included in the original tool code and are
planned to be added in a future change.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>